Add feature flag service
