+++ Production/govuk_ios/LocalFeatureFlag/FeatureFlagProviderInterface.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+typealias FeatureFlagCallback = ([FeatureFlag]) -> Void
+
+protocol FeatureFlagProviderInterface {
+   func fetchFeatureFlags(_ completion: @escaping FeatureFlagCallback)
+}
